requirements.txt
absl-py==1.4.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==25.1.0
autopep8==2.0.2
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
Brotli==1.1.0
cachetools==5.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
executing==1.2.0
fastjsonschema==2.21.1
filelock==3.9.0
flatbuffers==23.3.3
fonttools==4.39.0
fqdn==1.5.1
fsspec==2025.3.0
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.4
iniconfig==2.1.0
inquirerpy==0.3.4
ipykernel==6.24.0
ipython==8.14.0
isoduration==20.11.0
jax==0.4.6
jedi==0.18.2
Jinja2==3.1.6
joblib==1.3.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==2.12.0rc1
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.1.2
mpmath==1.3.0
mutagen==1.47.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.5.6
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.3
pfzy==0.3.4
pickleshare==0.7.5
Pillow==9.4.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
platformdirs==3.1.0
pluggy==1.6.0
prometheus_client==0.21.1
prompt-toolkit==3.0.39
protobuf==4.22.1
psutil==5.9.5
psycopg2==2.9.11
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.22
pycryptodomex==3.23.0
Pygments==2.15.1
pyparsing==3.0.9
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.2.1
pytz==2022.7.1
pywin32==306
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==25.1.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.44
stack-data==0.6.2
sympy==1.13.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0rc1
tensorflow-estimator==2.12.0rc0
tensorflow-intel==2.12.0rc1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.18.1
threadpoolctl==3.4.0
tinycss2==1.4.0
torchsummary==1.5.1
tornado==6.3.2
tqdm==4.67.1
traitlets==5.9.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==2.3.0
virtualenv==20.20.0
wcwidth==0.2.6
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.2.3
wrapt==1.14.1
yt-dlp==2025.6.9
